Convert png-grab to png-cam.
The major change is to react to the button press.  The picture is put in
the mmc when the button is pressed.
* projects/png-cam/main.c (file_offset):  Static variable that gives us
a picture name counter.
(main): Fix an 'unused variable' warning.  Fix an 'obscured variable'
warning.  Make sure we take the picture when the button is pressed.
(get_next_file_name): Helper function to take out the logic of naming
the file out of the main function.