Allow a variable list of arguments for the DEBUG functions.
This adds a more general layer on top of the already existing debug
layer.  The only annoying thing is that you must always call the
debugging functions with a format string and a list args.  This is not
always optimal.
The idea is that, in the future, there might be other debugging types
that could fit into the system.  To add a new "debugging type" one just
adds it to the CC3_DEBUG_XXX variables.  One can also add a new DEBUG
macro to go with the var.
* hal/lpc2106-cmucam3/cc3_debug.c (cc3_debug_debug)
(cc3_debug_error):  Add code to handle variable arguments.
* include/cc3_debug.h (CC3_ERROR, CC3_PDEBUG, CC3_HWDEBUG):  Likewise.
* projects/png-grab/main.c (main, capture_png):  Use the new variable
argument code.